Modify MasqDaemon to allow multiple chains
MasqDaemon was designed to have only one chain since it masquerades
all traffic leaving the node. In my use case I want to use MasqDaemon
to manage multiple chains so that limitation was a problem.

Now the chain object is passed into the MasqDaemon so the masqChain
var is no longer needed.