Replace comment that's no longer relevant
I believe the original issue occurred because the whole (after "-m")
" comment --comment ..." string was parsed as the module name (in place
of "comment"). It's not possible to put unparsed string in the args,
because it doesn't go through a shell.

Also removed the unnecessary quote; I think it was there because we
assumed it's passing through the shell (so that words get split).